authorized condition in update & delete

+++ src/tasks/controller.py
@@ -23,8 +23,8 @@ def create_task(body: TaskSchema, db: Session, user:UserModel):
     return new_task
 
 
-def get_tasks(db: Session):
-    tasks = db.query(TaskModel).all()
+def get_tasks(db: Session, user:UserModel):
+    tasks = db.query(TaskModel).filter(TaskModel.user_id == user.id).all()
     return tasks
 
 
@@ -36,10 +36,13 @@ def get_one_task(task_id: int, db: Session):
     return one_task
 
 
-def update_task(body: TaskSchema, task_id: int, db: Session):
-    one_task = db.query(TaskModel).get(task_id)
+def update_task(body: TaskSchema, task_id: int, db: Session, user:UserModel):
+    one_task:TaskModel = db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404, detail="Task Not Found")
+    
+    if not one_task.user_id != user.id:
+        raise HTTPException(401,detail='You are not allowed to update this task')
 
     # only specific details
     body = body.model_dump()
@@ -59,10 +62,13 @@ def update_task(body: TaskSchema, task_id: int, db: Session):
     return one_task
 
 
-def delete_task(task_id: int, db: Session):
+def delete_task(task_id: int, db: Session, user:UserModel):
     one_task = db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404, detail="Task Not Found")
+    
+    if not one_task.user_id != user.id:
+        raise HTTPException(401,detail='You are not allowed to update this task')
 
     db.delete(one_task)
     db.commit()
